finish order voucher and report

+++ src/main/java/org/weebook/api/service/impl/VoucherServiceImpl.java
@@ -25,11 +25,17 @@ import org.weebook.api.repository.UserRepository;
 import org.weebook.api.repository.VoucherRepository;
 import org.weebook.api.service.VoucherService;
 import org.weebook.api.util.CriteriaUtil;
+import org.weebook.api.util.ExecutorUtils;
 import org.weebook.api.web.request.AddVoucherVaoUserRequest;
 import org.weebook.api.web.request.FilterRequest;
 import org.weebook.api.web.request.PagingRequest;
 import org.weebook.api.web.request.VoucherRequest;
 
+import java.math.BigDecimal;
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.util.List;
 
 @Service
@@ -50,15 +56,23 @@ public class VoucherServiceImpl implements VoucherService {
 
     @Override
     public VoucherDTO create(VoucherRequest voucherRequest) {
-        Specification<User> specification = getSpecificationOr(voucherRequest.getFilterRequest());
-        List<User> users = userRepository.findAll(specification);
         Voucher voucher = voucherMapper.requestToEntity(voucherRequest);
-        if (voucherRequest.getFilterRequest() == null) {
-            userAddNotifications(users, TITLE, "Bạn có 1 voucher mới :" + voucher.getCode(), TYPE);
-            voucherRepository.save(voucher);
-        } else {
-            userAddNotificationsAndVoucher(users, voucher, TITLE, "Bạn có 1 voucher mới :" + voucher.getCode(), TYPE);
+        if(voucherRequest.getType().equals("%")){
+            if(voucherRequest.getDiscountAmount().compareTo(BigDecimal.valueOf(100)) >0){
+                throw new StringException("Giảm phần % thì nhập số nhỏ hơn 100");
+            }
         }
+        Runnable task = () -> {
+            if (voucherRequest.getFilterRequest() == null) {
+                notification(TITLE, "Bạn có 1 voucher mới :" + voucher.getCode(), TYPE);
+                voucherRepository.save(voucher);
+            } else {
+                Specification<User> specification = getSpecificationOr(voucherRequest.getFilterRequest());
+                List<User> users = userRepository.findAll(specification);
+                userVoucherNotification(users, voucher, TITLE, "Bạn có 1 voucher mới :" + voucher.getCode(), TYPE);
+            }
+        };
+        ExecutorUtils.executor.execute(task);
         return voucherMapper.entityToDto(voucher);
     }
 
@@ -71,28 +85,29 @@ public class VoucherServiceImpl implements VoucherService {
         if (vouchers.isEmpty()) {
             return null;
         }
-
         Voucher checkVoucher = vouchers.get(0);
-        if (checkVoucher.getUser() == null) {
-            List<User> users = userRepository.findAll();
-            String message = "Xin loi vì voucher : " + addVoucherVaoUserRequest.getVoucherDTO().getCode() + " chỉ dành cho những khách đặc biệt";
-            userAddNotifications(users, "Voucher", message, "voucher");
-            voucherRepository.deleteVoucherByCodeEquals(addVoucherVaoUserRequest
-                    .getVoucherDTO().getCode());
-        }
 
-        Specification<User> specification = getSpecificationOr(addVoucherVaoUserRequest.getFilterRequest());
-        List<User> users = userRepository.findAll(specification);
-
-        if (addVoucherVaoUserRequest.getFilterRequest() == null) {
-            throw new StringException("Lần trước nhậu quên nhập giờ thì nhớ nha.");
-        }
-
-        Voucher voucher = voucherMapper.dtoToEntity(addVoucherVaoUserRequest.getVoucherDTO());
-
-        userAddNotificationsAndVoucher(users, voucher, TITLE, "Bạn có 1 voucher mới :" + voucher.getCode(), TYPE);
+                Specification<User> specification = getSpecificationOr(addVoucherVaoUserRequest.getFilterRequest());
+                List<User> userSpec = userRepository.findAll(specification);
+                if (checkVoucher.getUser() == null) {
+                    String message = "Xin loi vì voucher : " + addVoucherVaoUserRequest.getVoucherDTO().getCode() + " chỉ dành cho những khách đặc biệt";
+                    subtract_notification(userSpec,"Delete Voucher", message, TYPE);
+                    voucherRepository.deleteVoucher(addVoucherVaoUserRequest.getVoucherDTO().getCode());
+                    userVoucher(userSpec, checkVoucher);
+                }else{
+                    List<User> userOnVoucher = voucherRepository.findByVoucherCode(addVoucherVaoUserRequest.getVoucherDTO().getCode());
+                    List<User> users = userRepository.getUser(userSpec, userOnVoucher);
+                    userVoucherNotification(users, checkVoucher, TITLE, "Bạn có 1 voucher mới :" + checkVoucher.getCode(), TYPE);
+                }
 
         return addVoucherVaoUserRequest.getVoucherDTO();
+    }
+
+    @Override
+    public void update(VoucherRequest voucherRequest) {
+        Instant validTo = instant(voucherRequest.getValidTo());
+        Instant validFrom = instant(voucherRequest.getValidFrom());
+        voucherRepository.updateVoucher(voucherRequest.getType(), voucherRequest.getDescription(), validTo, validFrom, voucherRequest.getDiscountAmount(), voucherRequest.getCondition(), voucherRequest.getCode());
     }
 
     @Override
@@ -104,7 +119,6 @@ public class VoucherServiceImpl implements VoucherService {
 
         String userName = currentUser.getName();
         User user = (User) userDetailsService.loadUserByUsername(userName);
-        System.out.println(user.getId());
         List<Voucher> vouchers = voucherRepository.userGetAll(user.getId());
         return voucherMapper.entityToDtos(vouchers);
     }
@@ -119,10 +133,23 @@ public class VoucherServiceImpl implements VoucherService {
     @Transactional
     @Override
     public String delete(String code) {
-        List<User> users = voucherRepository.findByVoucherCode(code);
-        userAddNotifications(users, TITLE, MESSAGE_DELETE + code, TYPE);
-        userRepository.saveAllAndFlush(users);
-        voucherRepository.deleteVoucherByCodeEquals(code);
+        List<Voucher> vouchers = voucherRepository.findByCodeEquals(code);
+        if (vouchers.isEmpty()) {
+            return null;
+        }
+        Voucher checkVoucher = vouchers.get(0);
+        Runnable task = () -> {
+            if (checkVoucher.getUser() == null) {
+                String message = "Xin lỗi vì đã xóa voucher : " + code;
+                notification("Delete Voucher", message, TYPE);
+            }else{
+                List<User> users = voucherRepository.findByVoucherCode(code);
+                userAddNotifications(users, TITLE, MESSAGE_DELETE + code, TYPE);
+                userRepository.saveAllAndFlush(users);
+            }
+            voucherRepository.deleteVoucherByCodeEquals(code);
+        };
+        ExecutorUtils.executor.execute(task);
         return "Delete successfully";
     }
 
@@ -137,24 +164,54 @@ public class VoucherServiceImpl implements VoucherService {
             return;
         }
         for (User user : users) {
-            Notification notification = notificationMapper.notification(title, message, type, user);
-            notification = notificationRepository.save(notification);
-            user.getNotifications().add(notification);
+            user.getNotifications().add(notification(title, message, type, user));
         }
     }
 
-    public void userAddNotificationsAndVoucher(List<User> users, Voucher voucher, String title, String message, String type) {
+    public Notification notification(String title, String message, String type, User user){
+        Notification notification = notificationMapper.notification(title,message,type, user);
+        return notificationRepository.save(notification);
+    }
+
+    public Notification notification(String title, String message, String type){
+        Notification notification = notificationMapper.notification(title,message,type);
+        return notificationRepository.save(notification);
+    }
+
+    public void subtract_notification(List<User> users, String title, String message, String type){
+        Notification notification = notificationMapper.notification(title,message,type);
+        notification.setUsers(users);
+        notificationRepository.save(notification);
+    }
+
+    public void userVoucherNotification(List<User> users, Voucher voucher, String title, String message, String type) {
         if (users.isEmpty()) {
             return;
         }
         for (User user : users) {
-            Notification notification = notificationMapper.notification(title, message, type, user);
-            notification = notificationRepository.save(notification);
-            user.getNotifications().add(notification);
+            Notification notification = notificationMapper.notification(title,message,type);
+            notification.setUser(user);
+            notificationRepository.save(notification);
+            System.out.println("asdffsad");
 
             Voucher voucherNew = voucherMapper.entityToEntity(voucher, user);
-            voucherNew = voucherRepository.save(voucherNew);
-            user.getVouchers().add(voucherNew);
+            voucherRepository.save(voucherNew);
+
+
+        }
+    }
+
+    public Voucher voucher(User user, Voucher voucher){
+        Voucher voucherNew = voucherMapper.entityToEntity(voucher, user);
+        return voucherRepository.save(voucherNew);
+    }
+
+    public void userVoucher(List<User> users, Voucher voucher) {
+        if (users.isEmpty()) {
+            return;
+        }
+        for (User user : users) {
+            user.getVouchers().add(voucher(user, voucher));
         }
     }
 
@@ -166,6 +223,11 @@ public class VoucherServiceImpl implements VoucherService {
         return filterRequests.stream()
                 .<Specification<T>>map(CriteriaUtil::toSpecification)
                 .reduce(Specification.where(null), Specification::or);
+    }
+
+    Instant instant(LocalDateTime localDateTime) {
+        ZonedDateTime zonedDateTime = localDateTime.atZone(ZoneId.systemDefault());
+        return zonedDateTime.toInstant();
     }
